+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- Prompt to fetch new content appears when a new topic is made in the active board
+
 ## [0.8.1] - 2021-10-08
 
 ### Fixed
@@ -203,6 +209,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User cache for reduced DB queries
 - Page title updates to reflect current view
 
+[Unreleased]: https://github.com/APB9785/phxBB/compare/0.8.1...HEAD
 [0.8.1]: https://github.com/APB9785/phxBB/compare/0.8...0.8.1
 [0.8]: https://github.com/APB9785/phxBB/compare/0.7.1...0.8
 [0.7.1]: https://github.com/APB9785/phxBB/compare/0.7...0.7.1